requirements.txt
fastapi
uvicorn[standard]
kafka-python
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
streamlit
pandas
plotly